+++ habitus.api/Controllers/HabitsController.cs
@@ -8,14 +8,17 @@ namespace habitus.api.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    [Authorize(Policy = "SameUser")]
+    [Authorize]
     public class HabitsController : ControllerBase
     {
         private readonly IRepositoryManager _repository;
+        private readonly IAuthorizationService _authorization;
 
-        public HabitsController(IRepositoryManager repository)
+        public HabitsController(IRepositoryManager repository, IAuthorizationService authorizationService)
         {
             _repository = repository;
+            _authorization = authorizationService;
+
         }
 
         // GET: api/Habits
@@ -23,17 +26,17 @@ namespace habitus.api.Controllers
         public async Task<ActionResult<IEnumerable<HabitResponse>>> Get()
         {
             string userId = GetUserId();
-
             try
             {
                 IEnumerable<HabitResponse> habits = await _repository.Habit.FindAllHabits(userId);
 
-                if (habits == null)
-                {
-                    return NotFound();
-                }
+                if (habits == null) return NotFound();
 
-                return Ok(habits);
+                var authorizationResult = await _authorization.AuthorizeAsync(User, habits.First(), "SameUser");
+
+                if (authorizationResult.Succeeded) return Ok(habits);
+                else if (User.Identity!.IsAuthenticated) return new ForbidResult();
+                else return new ChallengeResult();
             }
             catch (Exception ex)
             {
@@ -44,18 +47,17 @@ namespace habitus.api.Controllers
         [HttpGet("{id}")]
         public async Task<ActionResult<IEnumerable<HabitResponse>>> Get(int id)
         {
-            string userId = GetUserId();
-
             try
             {
                 var habit = await _repository.Habit.Find(id);
 
-                if (habit == null)
-                {
-                    return NotFound();
-                }
+                if (habit == null) return NotFound();
 
-                return Ok(habit);
+                var authorizationResult = await _authorization.AuthorizeAsync(User, habit, "SameUser");
+
+                if (authorizationResult.Succeeded) return Ok(habit);
+                else if (User.Identity!.IsAuthenticated) return new ForbidResult();
+                else return new ChallengeResult();
             }
             catch (Exception ex)
             {
@@ -76,10 +78,12 @@ namespace habitus.api.Controllers
             {
                 var habits = await _repository.Habit.FindAllAndFilterEntriesByDate(filter.StartDate, filter.EndDate.Value, userId);
 
-                if (habits == null)
-                {
-                    return NotFound();
-                }
+                if (habits == null) return NotFound();
+
+                var authorizationResult = await _authorization.AuthorizeAsync(User, habits.First(), "SameUser");
+
+                if (!authorizationResult.Succeeded && User.Identity!.IsAuthenticated) return new ForbidResult();
+                else if (!authorizationResult.Succeeded) return new ChallengeResult();
 
                 return Ok(habits);
             }
@@ -93,11 +97,14 @@ namespace habitus.api.Controllers
         [HttpPost]
         public async Task<IActionResult> Post(CreateHabitRequest request)
         {
-            string userId = GetUserId();
-
             try
             {
-                int id = await _repository.Habit.CreateHabit(request, userId);
+                var authorizationResult = await _authorization.AuthorizeAsync(User, request, "SameUser");
+
+                if (!authorizationResult.Succeeded && User.Identity!.IsAuthenticated) return new ForbidResult();
+                else if (!authorizationResult.Succeeded) return new ChallengeResult();
+
+                int id = await _repository.Habit.CreateHabit(request);
 
                 switch (id)
                 {
@@ -120,19 +127,23 @@ namespace habitus.api.Controllers
         {
             try
             {
+                var habit = await _repository.Habit.Find(id);
+
+                if (habit == null) return NotFound();
+
+                var authorizationResult = await _authorization.AuthorizeAsync(User, habit, "SameUser");
+
+                if (!authorizationResult.Succeeded && User.Identity!.IsAuthenticated) return new ForbidResult();
+                else if (!authorizationResult.Succeeded) return new ChallengeResult();
+
                 bool response = await _repository.Habit.DeleteHabit(id);
-
-                if (!response)
-                {
-                    return NotFound();
-                }
-
-                return NoContent();
             }
             catch (Exception ex)
             {
                 return Problem(ex.Message);
             }
+
+            return NoContent();
         }
 
         // PUT: api/Habits/5
@@ -143,12 +154,18 @@ namespace habitus.api.Controllers
 
             try
             {
+                var habit = await _repository.Habit.Find(id);
+
+                if (habit == null) return NotFound();
+
+                var authorizationResult = await _authorization.AuthorizeAsync(User, habit, "SameUser");
+
+                if (!authorizationResult.Succeeded && User.Identity!.IsAuthenticated) return new ForbidResult();
+                else if (!authorizationResult.Succeeded) return new ChallengeResult();
+
                 bool response = await _repository.Habit.UpdateHabit(request);
 
-                if (!response)
-                {
-                    return NotFound();
-                }
+                if (!response) return NotFound();
             }
             catch (DbUpdateConcurrencyException)
             {
@@ -157,7 +174,6 @@ namespace habitus.api.Controllers
                     return NotFound();
                 }
             }
-            // NOTE I have not investigated if this ais a proper way to chain exception catching
             catch (Exception ex)
             {
                 return Problem(ex.Message);
